requirements.txt
requests
pandas
langchain-google-genai
matplotlib
pyarrow
streamlit
torch
sentence-transformers
numpy
faiss-cpu